--- service/src/main/scala/com/namely/chiefofstate/Constants.scala
@@ -1,8 +1,0 @@
-package com.namely.chiefofstate
-
-object Constants {
-
-  val READ_SIDE_HOST_KEY: String = "HOST"
-  val READ_SIDE_PORT_KEY: String = "PORT"
-
-}
