+++ service/src/test/scala/com/namely/chiefofstate/ReadSideHandlerSpec.scala
@@ -18,7 +18,7 @@ import com.namely.protobuf.chief_of_state.readside.{
 import com.namely.protobuf.chief_of_state.tests.{Account, AccountOpened}
 import com.namely.chiefofstate.config.{
   HandlerSetting,
-  ReadSideConfig
+  ReadSideSetting
 }
 import io.grpc.Status
 import io.superflat.lagompb.protobuf.core.MetaData
@@ -51,7 +51,7 @@ class ReadSideHandlerSpec
     with MockFactory {
 
   val sys: ActorSystem[Nothing] = testKit.system
-  val defaultGrpcReadSideConfig: ReadSideConfig = ReadSideConfig("test")
+  val defaultGrpcReadSideSetting: ReadSideSetting = ReadSideSetting("test")
 
   "Chief-Of-State ReadSide Processor" should {
 
@@ -125,7 +125,7 @@ class ReadSideHandlerSpec
 
       val readSideProcessor =
         new ReadSideHandler(
-          defaultGrpcReadSideConfig,
+          defaultGrpcReadSideSetting,
           encryptionAdapter,
           testKit.system.toClassic,
           mockGrpcClient,
@@ -214,7 +214,7 @@ class ReadSideHandlerSpec
 
       val readSideProcessor =
         new ReadSideHandler(
-          defaultGrpcReadSideConfig,
+          defaultGrpcReadSideSetting,
           encryptionAdapter,
           testKit.system.toClassic,
           mockGrpcClient,
@@ -296,7 +296,7 @@ class ReadSideHandlerSpec
 
       val readSideProcessor =
         new ReadSideHandler(
-          defaultGrpcReadSideConfig,
+          defaultGrpcReadSideSetting,
           encryptionAdapter,
           testKit.system.toClassic,
           mockGrpcClient,
@@ -382,7 +382,7 @@ class ReadSideHandlerSpec
 
       val readSideProcessor =
         new ReadSideHandler(
-          defaultGrpcReadSideConfig,
+          defaultGrpcReadSideSetting,
           encryptionAdapter,
           testKit.system.toClassic,
           mockGrpcClient,
@@ -468,7 +468,7 @@ class ReadSideHandlerSpec
 
       val readSideProcessor =
         new ReadSideHandler(
-          defaultGrpcReadSideConfig,
+          defaultGrpcReadSideSetting,
           encryptionAdapter,
           testKit.system.toClassic,
           mockGrpcClient,
@@ -505,7 +505,7 @@ class ReadSideHandlerSpec
 
       val readSideProcessor =
         new ReadSideHandler(
-          defaultGrpcReadSideConfig,
+          defaultGrpcReadSideSetting,
           encryptionAdapter,
           testKit.system.toClassic,
           null,
